requirements.txt
Django==3.2.0
psycopg2-binary==2.9.3
djangorestframework==3.13.1